Add files via upload
This is the job analysis engine script. Runs all pertinent analyses and outputs an eventual XLSX workbook document outlining the pertinent findings of the study. The code is highly variable to accept all the different types of data and scales consultants are able to implement during these types of studies.